disabled "Experimental"
The "Experimental" option has been disabled in the UI dropdown
list for selecting 'Voices' due to it causing problems: note C4
stops working after trying to play a note using the "Experimental"
voice.

This is the last commit before uploading updates to the live app'
to make "The Voice of chatGPT" available to users.